Attempt to explain why a certain Bioconductor package fails on the Bioc build machine(s)

Attempt to explain why a certain Bioconductor package fails on the Bioc build machine(s)

explainBiocBuildFailure(
  package,
  biocVersion = BiocManager::version(),
  buildNodes = c("nebbiolo2", "tokay2", "machv2", "nebbiolo1", "riesling1",
    "palomino3", "merida1"),
  showOnlyFailing = FALSE,
  repo = "BioCsoft",
  dependencies = c("Depends", "Imports")
)

Arguments

package

Character string giving the package of interest.

biocVersion

Character string giving the Bioconductor version.

buildNodes

Character vector indicating the build nodes to check. Should be a subset of c("nebbiolo2", "tokay2", "machv2", "nebbiolo1", "riesling1", "palomino3", "merida1"). If more than one node is specified, a package is considered to fail if it fails on at least one of them. A failure is recognized by an "ERROR" returned in the build report, in any of the stages.

showOnlyFailing

Logical, whether to show all dependencies or just the failing ones (and the intermediate nodes between the failing ones and the package of interest) in the output graph.

repo

Character vector indicating the repo(s) to consider. Will be passed to biocPkgList().

dependencies

Character vector indicating the types of dependencies to consider. Will be passed to biocPkgList().

Value

Plots a graph containing the package of interest as well as its dependencies. For each package, the graph displays the name, the date when it was last changed, and the fraction of its direct descendants (packages that depend on or import it) that currently fail. Note that only Bioc dependencies are shown.

Author

Charlotte Soneson

Examples

explainBiocBuildFailure("Rhisat2")
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org